+++ packages/frameworks/angular/src/password-input/password-input-visibility-trigger.directive.ts
@@ -6,6 +6,7 @@ import {Eye, EyeOff} from "lucide-angular"
 
 import {useInlineIconButtonApi} from "@qualcomm-ui/angular/inline-icon-button"
 import {useQdsInputContext} from "@qualcomm-ui/angular/input"
+import {QuiPreloadDirective} from "@qualcomm-ui/angular/transitions"
 import {
   type LucideIconOrString,
   provideIcons,
@@ -15,6 +16,7 @@ import {passwordInputClasses} from "@qualcomm-ui/qds-core/password-input"
 import {mergeProps} from "@qualcomm-ui/utils/merge-props"
 
 @Component({
+  hostDirectives: [QuiPreloadDirective],
   providers: [provideIcons({Eye, EyeOff})],
   selector: "[q-password-input-visibility-trigger]",
   standalone: false,