+++ packages/frameworks/angular/src/switch/switch-thumb.directive.ts
@@ -3,11 +3,13 @@
 
 import {computed, Directive} from "@angular/core"
 
+import {QuiPreloadDirective} from "@qualcomm-ui/angular/transitions"
 import {CoreSwitchThumbDirective} from "@qualcomm-ui/angular-core/switch"
 
 import {useQdsSwitchContext} from "./qds-switch-context.service"
 
 @Directive({
+  hostDirectives: [QuiPreloadDirective],
   selector: "[q-switch-thumb]",
   standalone: false,
 })